change option -A to -A channel_id and add a new template

+++ openair1/PHY/impl_defs_top.h
@@ -322,7 +322,7 @@ typedef struct
   // RRC measurements
   int rssi;
   int n_adj_cells;
-  u16 adj_cell_id[6];
+  unsigned int adj_cell_id[6];
   int rsrq[7];
   int rsrp[7];
   // common measurements
